backend/requirements.txt
fastapi
uvicorn
sqlmodel
psycopg2-binary
python-dotenv
pyjwt
passlib[bcrypt]
bcrypt>=3.2,<4.0
pytest
httpx
pytest-cov